修复: packmoney和green_discount全面支持小数(int→double，去掉Math.floor截断)，PHP侧packmoney强转int改为float

+++ src/main/java/Sdf1_login/ConfigManager.java
@@ -60,8 +60,8 @@ public class ConfigManager {
     public int tpSendIntervalSeconds = 10; // 发送间隔（秒）
 
     // ===== 商店打包配置（来自 PHP shop_config，定时拉取）=====
-    public int packingFee = 5;        // 彩色潜影盒打包加收债券数（原色免费）
-    public int greenDiscount = 2;     // 环保单减免百分比（不打包/塞背包时减免，0=不减免）
+    public double packingFee = 5.0;     // 彩色潜影盒打包加收债券数（原色免费，支持小数）
+    public double greenDiscount = 2.0;   // 环保单减免百分比（不打包/塞背包时减免，0=不减免，支持小数）
 
     // ★ 构造函数接收 File，和 Main.java 调用一致 ★
     public ConfigManager(File dataFolder) {
